data_supply/requirements.txt
websocket-client>=0.58.0
psutil>=5.8.0